Fix build for new Heimdal, new Config::AutoConf
Current versions of Heimdal no longer include all of the libraries
when asking pkg-config for heimdal-kadm-server, so explicitly ask
for heimdal-krb5 as well and concatenate the flags together.

Pass the compiler flags into Config::AutoConf as well as the linker
flags to pick up Heimdal -I/usr/include/heimdal flags on Debian.

Adjust the call to check_decl for the new Config::AutoConf call
style.